Switched the ChunkRelativePos struct to use u8 for storage instead of usize, same with the chunk size constant

+++ src/engine/common.rs
@@ -1,10 +1,10 @@
 pub struct ChunkRelativePos {
-    pub x: usize,
-    pub y: usize,
+    pub x: u8,
+    pub y: u8,
 }
 
 impl ChunkRelativePos {
-    pub fn new(x: usize, y: usize) -> ChunkRelativePos {
+    pub fn new(x: u8, y: u8) -> ChunkRelativePos {
         return ChunkRelativePos {
             x,
             y,
